split reducers
