+++ project/src/middleware/thunk-api.js
@@ -1,36 +1,59 @@
 import {ApiRoute, AppRoute} from '../const.js';
-import {ActionCreator} from '../store/action.js';
+import {fetchFilmsFail, fetchFilmsStart, fetchFilmsSuccess} from '../store/films/films-action.js';
+import {fetchFavoritesFail, fetchFavoritesStart, fetchFavoritesSuccess} from '../store/favorites/favorites-action.js';
+import {redirectToRoute} from '../store/process/process-action.js';
+import {
+  authorizationFail,
+  authorizationSuccess,
+  logoutUserFail,
+  logoutUserSuccess,
+  sendAuthorizationRequest,
+  setUserLoginProfile
+} from '../store/user/user-action.js';
+import {
+  fetchSimilarFilmsFail,
+  fetchSimilarFilmsStart,
+  fetchSimilarFilmsSuccess
+} from '../store/similar-films/similar-films-action.js';
+import {
+  addReviewFail,
+  addReviewStart,
+  addReviewSuccess,
+  fetchReviewsFail,
+  fetchReviewsStart,
+  fetchReviewsSuccess
+} from '../store/reviews/reviews-action.js';
 
 export const fetchFilms = () => (dispatch, _getState, api) => {
-  dispatch(ActionCreator.fetchFilmsStart());
+  dispatch(fetchFilmsStart());
   return api
     .get(ApiRoute.FETCH_FILMS)
-    .then(({data}) => dispatch(ActionCreator.fetchFilmsSuccess(data)))
-    .catch((err) => dispatch(ActionCreator.fetchFilmsFail(err)));
+    .then(({data}) => dispatch(fetchFilmsSuccess(data)))
+    .catch((err) => dispatch(fetchFilmsFail(err)));
 };
 
 export const checkAuthUser = () => (dispatch, _getState, api) => {
-  dispatch(ActionCreator.sendAuthorizationRequest());
+  dispatch(sendAuthorizationRequest());
   return api
     .get(ApiRoute.LOGIN)
-    .then(({data}) => dispatch(ActionCreator.setUserLoginProfile(data)))
-    .then(() => dispatch(ActionCreator.authorizationSuccess()))
-    .catch((err) => dispatch(ActionCreator.authorizationFail(err)));
+    .then(({data}) => dispatch(setUserLoginProfile(data)))
+    .then(() => dispatch(authorizationSuccess()))
+    .catch((err) => dispatch(authorizationFail(err)));
 };
 
 export const loginUser = ({email, password}) => (dispatch, _getState, api) => {
-  dispatch(ActionCreator.sendAuthorizationRequest());
+  dispatch(sendAuthorizationRequest());
   return api
     .post(ApiRoute.LOGIN, {email, password})
     .then(({data}) => {
-      dispatch(ActionCreator.setUserLoginProfile(data));
+      dispatch(setUserLoginProfile(data));
       localStorage.setItem('token', data.token);
     })
-    .then(() => dispatch(ActionCreator.authorizationSuccess()))
-    .then(() => dispatch(ActionCreator.redirectToRoute(AppRoute.ROOT)))
+    .then(() => dispatch(authorizationSuccess()))
+    .then(() => dispatch(redirectToRoute(AppRoute.ROOT)))
     .catch((err) => {
-      dispatch(ActionCreator.authorizationFail(err));
-      dispatch(ActionCreator.redirectToRoute(AppRoute.LOGIN));
+      dispatch(authorizationFail(err));
+      dispatch(redirectToRoute(AppRoute.LOGIN));
     });
 };
 
@@ -39,44 +62,44 @@ export const logoutUser = () => (dispatch, _getState, api) => (
     .delete(ApiRoute.LOGOUT)
     .then(() => {
       localStorage.removeItem('token');
-      dispatch(ActionCreator.logoutUserSuccess());
+      dispatch(logoutUserSuccess());
     })
-    .catch((err) => dispatch(ActionCreator.logoutUserFail(err)))
+    .catch((err) => dispatch(logoutUserFail(err)))
 );
 
 export const fetchFavorites = () => (dispatch, _getState, api) => {
-  dispatch(ActionCreator.fetchFavoritesStart());
+  dispatch(fetchFavoritesStart());
   return api
     .get(ApiRoute.FETCH_FAVORITES)
-    .then(({data}) => dispatch(ActionCreator.fetchFavoritesSuccess(data)))
-    .catch((err) => dispatch(ActionCreator.fetchFavoritesFail(err)));
+    .then(({data}) => dispatch(fetchFavoritesSuccess(data)))
+    .catch((err) => dispatch(fetchFavoritesFail(err)));
 };
 
 export const fetchSimilarFilms = (id) => (dispatch, _getState, api) => {
-  dispatch(ActionCreator.fetchSimilarFilmsStart());
+  dispatch(fetchSimilarFilmsStart());
   return api
     .get(ApiRoute.FETCH_SIMILAR_FILMS.replace(':id', id))
-    .then(({data}) => dispatch(ActionCreator.fetchSimilarFilmsSuccess(data)))
-    .catch((err) => dispatch(ActionCreator.fetchSimilarFilmsFail(err)));
+    .then(({data}) => dispatch(fetchSimilarFilmsSuccess(data)))
+    .catch((err) => dispatch(fetchSimilarFilmsFail(err)));
 };
 
 export const fetchReviews = (id) => (dispatch, _getState, api) => {
-  dispatch(ActionCreator.fetchReviewsStart());
+  dispatch(fetchReviewsStart());
   return api
     .get(ApiRoute.FETCH_REVIEWS.replace(':film_id', id))
-    .then(({data}) => dispatch(ActionCreator.fetchReviewsSuccess(data)))
-    .catch((err) => dispatch(ActionCreator.fetchReviewsFail(err)));
+    .then(({data}) => dispatch(fetchReviewsSuccess(data)))
+    .catch((err) => dispatch(fetchReviewsFail(err)));
 };
 
 export const addReview = (id, {rating, comment}) => (dispatch, _getState, api) => {
-  dispatch(ActionCreator.addReviewStart());
+  dispatch(addReviewStart());
   return api
     .post(ApiRoute.ADD_REVIEW.replace(':film_id', id), {rating, comment})
-    .then(({data}) => dispatch(ActionCreator.fetchReviewsSuccess(data)))
-    .then(() => dispatch(ActionCreator.addReviewSuccess()))
-    .then(() => dispatch(ActionCreator.redirectToRoute(AppRoute.FILM.replace(':id', id))))
+    .then(({data}) => dispatch(fetchReviewsSuccess(data)))
+    .then(() => dispatch(addReviewSuccess()))
+    .then(() => dispatch(redirectToRoute(AppRoute.FILM.replace(':id', id))))
     .catch((err) => {
-      dispatch(ActionCreator.addReviewFail(err));
-      dispatch(ActionCreator.redirectToRoute(AppRoute.REVIEW.replace(':id', id)));
+      dispatch(addReviewFail(err));
+      dispatch(redirectToRoute(AppRoute.REVIEW.replace(':id', id)));
     });
 };
