+++ project/src/store/action-type.js
@@ -1,0 +1,27 @@
+export const ActionType = {
+  AUTHORIZATION_SUCCESS: 'AUTHORIZATION_SUCCESS',
+  AUTHORIZATION_FAIL: 'AUTHORIZATION_FAIL',
+  CHANGE_CURRENT_GENRE: 'CHANGE_CURRENT_GENRE',
+  FETCH_REVIEWS_START: 'FETCH_REVIEWS_START',
+  FETCH_REVIEWS_SUCCESS: 'FETCH_REVIEWS_SUCCESS',
+  FETCH_REVIEWS_FAIL: 'FETCH_REVIEWS_FAIL',
+  ADD_REVIEW_START: 'ADD_REVIEW_START',
+  ADD_REVIEW_SUCCESS: 'ADD_REVIEW_SUCCESS',
+  ADD_REVIEW_FAIL: 'ADD_REVIEW_FAIL',
+  LOGOUT_USER_SUCCESS: 'LOGOUT_USER_SUCCESS',
+  LOGOUT_USER_FAIL: 'LOGOUT_USER_FAIL',
+  FETCH_FAVORITES_START: 'FETCH_FAVORITES_START',
+  FETCH_FAVORITES_SUCCESS: 'FETCH_FAVORITES_SUCCESS',
+  FETCH_FAVORITES_FAIL: 'FETCH_FAVORITES_FAIL',
+  FETCH_FILMS_START: 'FETCH_FILMS_START',
+  FETCH_FILMS_SUCCESS: 'FETCH_FILMS_SUCCESS',
+  FETCH_FILMS_FAIL: 'FETCH_FILMS_FAIL',
+  FETCH_SIMILAR_FILMS_START: 'FETCH_SIMILAR_FILMS_START',
+  FETCH_SIMILAR_FILMS_SUCCESS: 'FETCH_SIMILAR_FILMS_SUCCESS',
+  FETCH_SIMILAR_FILMS_FAIL: 'FETCH_SIMILAR_FILMS_FAIL',
+  SEND_AUTHORIZATION_REQUEST: 'SEND_AUTHORIZATION_REQUEST',
+  SET_AUTHORIZATION_STATUS: 'SET_AUTHORIZATION_STATUS',
+  SET_USER_LOGIN_PROFILE: 'SET_USER_LOGIN_PROFILE',
+  REDIRECT_TO_ROUTE: 'REDIRECT_TO_ROUTE',
+};
+
